Simplify benchmark by replacing Playwright test with direct measurement module (STU-1290)
