+++ scripts/benchmark-site-editor/package.json
@@ -11,6 +11,7 @@
 	"dependencies": {
 		"@wp-playground/cli": "^3.0.22",
 		"chalk": "^5.3.0",
+		"playwright": "^1.58.1",
 		"tsx": "^4.7.0"
 	},
 	"devDependencies": {
